TS-988 rename application
