+++ libs/application/template-api-modules/src/lib/modules/templates/index.ts
@@ -128,8 +128,8 @@ import { RequestInspectionTemplateModule } from './aosh/request-inspection/reque
 import { RequestInspectionTemplateService } from './aosh/request-inspection/request-inspection.service'
 import { HealthInsuranceDeclarationModule } from './health-insurance-declaration/health-insurance-declaration.module'
 import { HealthInsuranceDeclarationService } from './health-insurance-declaration/health-insurance-declaration.service'
-import { PrimarySchoolModule } from './primary-school/new-primary-school.module'
-import { PrimarySchoolService } from './primary-school/new-primary-school.service'
+import { PrimarySchoolModule } from './primary-school/primary-school.module'
+import { PrimarySchoolService } from './primary-school/primary-school.service'
 import { WorkAccidentNotificationTemplateModule } from './aosh/work-accident-notification/work-accident-notification.module'
 import { WorkAccidentNotificationTemplateService } from './aosh/work-accident-notification/work-accident-notification.service'
 import { IdCardModule } from './id-card/id-card.module'
