+++ libs/application/template-api-modules/src/lib/modules/templates/primary-school/primary-school.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common'
 import { SharedTemplateAPIModule } from '../../shared'
 import { NationalRegistryXRoadModule } from '@island.is/api/domains/national-registry-x-road'
-import { PrimarySchoolService } from './new-primary-school.service'
+import { PrimarySchoolService } from './primary-school.service'
 import { FriggClientModule } from '@island.is/clients/mms/frigg'
 
 @Module({
