+++ libs/application/templates/primary-school/README.md
@@ -2,17 +2,17 @@
 
 ## About
 
-This application template allows applicants to apply for a new primary school
+This application template allows applicants to apply for a primary school
 
 ## URLs
 
-/umsoknir/nyr-grunnskoli/
+/umsoknir/grunnskoli/
 
 ## Integrations
 
 - [Þjóðskrá](https://skra.is): To get the applicants information
 
-Run `nx test application-templates-new-primary-school` to execute the unit tests via [Jest](https://jestjs.io).
+Run `nx test application-templates-primary-school` to execute the unit tests via [Jest](https://jestjs.io).
 
 ## Code owners and maintainers
